[Lap] pc refactor

+++ src/main/java/vn/edu/bkis/repository/EnrollmentRepository.java
@@ -1,13 +1,12 @@
 package vn.edu.bkis.repository;
 
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import vn.edu.bkis.dto.NewestStudentDto;
 import vn.edu.bkis.model.Enrollment;
 
-import java.util.List;
-
-public interface EnrollmentRepository extends JpaRepository<Enrollment, Integer> {
+public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
   @Query(value = "SELECT u.username AS username, "
       + "        c.title AS titleCourse, "
       + "        en.enrolled_at AS enrolledAt, "
